requirements.txt
Django>=4.2.0
Pillow>=9.0.0
whitenoise>=6.0.0
gunicorn>=20.1.0 